Improve data types for Galileo navigation data

+++ src/core/system_parameters/Galileo_E1.h
@@ -103,7 +103,7 @@ const std::vector<std::pair<int32_t, int32_t>> PAGE_TYPE_bit({{1, 6}});
 /*Page 1 - Word type 1: Ephemeris (1/4)*/
 const std::vector<std::pair<int32_t, int32_t>> IOD_nav_1_bit({{7, 10}});
 const std::vector<std::pair<int32_t, int32_t>> T0E_1_bit({{17, 14}});
-const double t0e_1_LSB = 60;
+const int32_t t0e_1_LSB = 60;
 const std::vector<std::pair<int32_t, int32_t>> M0_1_bit({{31, 32}});
 const double M0_1_LSB = PI_TWO_N31;
 const std::vector<std::pair<int32_t, int32_t>> e_1_bit({{63, 32}});
@@ -151,7 +151,7 @@ const double C_ic_4_LSB = TWO_N29;
 const std::vector<std::pair<int32_t, int32_t>> C_is_4_bit({{39, 16}});
 const double C_is_4_LSB = TWO_N29;
 const std::vector<std::pair<int32_t, int32_t>> t0c_4_bit({{55, 14}});  //
-const double t0c_4_LSB = 60;
+const int32_t t0c_4_LSB = 60;
 const std::vector<std::pair<int32_t, int32_t>> af0_4_bit({{69, 31}});  //
 const double af0_4_LSB = TWO_N34;
 const std::vector<std::pair<int32_t, int32_t>> af1_4_bit({{100, 21}});  //
@@ -198,7 +198,7 @@ const std::vector<std::pair<int32_t, int32_t>> A1_6_bit({{39, 24}});
 const double A1_6_LSB = TWO_N50;
 const std::vector<std::pair<int32_t, int32_t>> Delta_tLS_6_bit({{63, 8}});
 const std::vector<std::pair<int32_t, int32_t>> t0t_6_bit({{71, 8}});
-const double t0t_6_LSB = 3600;
+const int32_t t0t_6_LSB = 3600;
 const std::vector<std::pair<int32_t, int32_t>> WNot_6_bit({{79, 8}});
 const std::vector<std::pair<int32_t, int32_t>> WN_LSF_6_bit({{87, 8}});
 const std::vector<std::pair<int32_t, int32_t>> DN_6_bit({{95, 3}});
@@ -294,7 +294,7 @@ const double A_0G_10_LSB = TWO_N35;
 const std::vector<std::pair<int32_t, int32_t>> A_1G_10_bit({{103, 12}});
 const double A_1G_10_LSB = TWO_N51;
 const std::vector<std::pair<int32_t, int32_t>> t_0G_10_bit({{115, 8}});
-const double t_0G_10_LSB = 3600;
+const int32_t t_0G_10_LSB = 3600;
 const std::vector<std::pair<int32_t, int32_t>> WN_0G_10_bit({{123, 6}});
